+++ test-app/resources/changed/post.js
@@ -1,0 +1,3 @@
+if(!(changed('name') && previous.name === undefined)) {
+    cancel('changed and/or previous are broken...');
+}