+++ HISTORY.md
@@ -1,5 +1,36 @@
 # History
 
+## 0.6.8
+
+ - Fixed CORS incorrectly requiring a referer header
+ - Added `dpd.once(name, fn)` function to execute a realtime handler exactly once
+ - Added `dpd.off(name, [fn])` function to remove a realtime handler
+ - Added `dpd.socketReady(fn)` function to listen for the built-in `connect` event
+ - Added `dpd.socket` property to provide direct access to socket.io.
+ - Fixed bug where username and password could be updated by an unauthenticated put.
+  - Usernames and passwords can be updated only by that user's session, a root session, or an internal request.
+
+## 0.6.7
+
+ - Added new data editor
+ - Fixed major bug where calling error() would not always cancel the request
+ - Fixed bug where PUT would fail without an error if you provided a query
+ - Changed root behavior - no longer ignores cancel() in events
+ - Fixed bugs preventing events from being `emit()`ed to users in certain connection states
+ - Fixed bug where boolean query values (?bool=true) were not treated as booleans
+ - Fixed unnecessary error when parsing JSON body
+ - Added more intelligent body parsing
+ - Added `changed()` method in collection events
+ - Added `previous` object in collection events
+ - Fixed `dpd showkey` prompt for missing keys.json
+
+## 0.6.6
+ 
+ - Added CORS support
+ - Exposed the server object to modules as `process.server`
+ - Fixed a rare bug where the first request after a login would not be authenticated
+ - Fixed minor bug when loading only node modules
+
 ## 0.6.5
  
  - Fixed `process.send` bug